# perf: [CAL-1928] Deduplicate eventTypes.getByViewer query