## What does this PR do?
I was digging into #9448 and noticed that there is duplication of ` eventTypes.getByViewer` queries as:

```
/eventTypes/getByViewer,getByViewer?batch=1&input=%7B%220%22%3A%7B%22json%22%3Anull%2C%22meta%22%3A%7B%22values%22%3A%5B%22undefined%22%5D%7D%7D%2C%221%22%3A%7B%22json%22%3A%7B%22filters%22%3A%7B%22teamIds%22%3Anull%2C%22userIds%22%3Anull%7D%7D%2C%22meta%22%3A%7B%22values%22%3A%7B%22filters.teamIds%22%3A%5B%22undefined%22%5D%2C%22filters.userIds%22%3A%5B%22undefined%22%5D%7D%7D%7D%7D
```

instead of:

```
/eventTypes/getByViewer?batch=1&input=%7B%220%22%3A%7B%22json%22%3Anull%2C%22meta%22%3A%7B%22values%22%3A%5B%22undefined%22%5D%7D%7D%7D
```

However, the response being returned is identical for both `getByViewer` calls. This was due to `{ filters: { teamIds: null, eventIds: null } }` passed to the `eventTypes.getByViewer` query at event-types page when no filters are applied.

It is important to note that `getByViewer` query is slow in itself and batching 2 duplicate calls slows it down further down in returning response.

## Type of change

<!-- Please delete bullets that are not relevant. -->

- Chore (refactoring code, technical debt, workflow improvements)

## How should this be tested?

I've tested this code using [k6](https://k6.io/docs/get-started/running-k6/) with the script below:

```js
import http from 'k6/http';

export const options = {
  scenarios: {
    my_scenario1: {
      executor: 'constant-arrival-rate',
      duration: '30s', // total duration
      preAllocatedVUs: 50, // to allocate runtime resources     preAll

      rate: 50, // number of constant iterations given `timeUnit`
      timeUnit: '1s',
    },
  },
};

export default function () {
  const headers = {
    "accept": "*/*",
    "accept-language": "en-GB,en-US;q=0.9,en;q=0.8",
    "cookie": "WZRK_G=f9f97e08a166411f89eaa03cb253fcbb; __clnds=bt9ek987gs5; next-auth.csrf-token=ef9046a5484d478ffdbb26c1c2f01d7a43e6c956f9ea067e0878684b1d3918df%7C310be88cb9d5e21f7c65485dff4fd2027c50be5b6825e983f71a97e73182f326; next-auth.callback-url=http%3A%2F%2Flocalhost%3A3000%2F; next-auth.session-token=eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2R0NNIn0..yjQlecK8O9TyXH_k.kKIA1xulL6Y-VT30j0NeMvvKHemsWocDLKWG0vtG5j3_mmoTfjLBrt69_YIBQzQIkSspPoQCmHjaVSd5o70Oj8AnT7LLjetu7e9SrPoN31glLrGeCdZoEsvELalzoa5i-hjHylQGYPnJi9exADW2Fh1gzSb_4SjKCu7HJO5Fo5HluCm5cX9HQgLQjOGy4_1KeLKN6jcZJrK9S1GfM0X4Gemyf9FfYF4WHQPnj3DY49HqPV2ByZwDjvOgOY-A9uIel107dU4B776ApCxZ5PVLRLP_CRsUdBxN4IRbVhutW_uHdv_Q7Fe8Px9KfGQG.oG889zFKcHDT-3zg2rfXsw",
  };
  // Before
  // http.get('http://localhost:3000/api/trpc/eventTypes/getByViewer,getByViewer?batch=1&input=%7B%220%22%3A%7B%22json%22%3Anull%2C%22meta%22%3A%7B%22values%22%3A%5B%22undefined%22%5D%7D%7D%2C%221%22%3A%7B%22json%22%3A%7B%22filters%22%3A%7B%22teamIds%22%3Anull%2C%22userIds%22%3Anull%7D%7D%2C%22meta%22%3A%7B%22values%22%3A%7B%22filters.teamIds%22%3A%5B%22undefined%22%5D%2C%22filters.userIds%22%3A%5B%22undefined%22%5D%7D%7D%7D%7D', { headers });

  // After
  http.get('http://localhost:3000/api/trpc/eventTypes/getByViewer?batch=1&input=%7B%220%22%3A%7B%22json%22%3Anull%2C%22meta%22%3A%7B%22values%22%3A%5B%22undefined%22%5D%7D%7D%7D', { headers });
}
```

Results:

Before:

```
     http_req_waiting...............: avg=3.1s     min=852.33ms med=3.15s   max=4.41s    p(90)=3.41s   p(95)=3.68s
     http_reqs......................: 473    15.340525/s
     iteration_duration.............: avg=3.15s    min=871.83ms med=3.18s   max=4.75s    p(90)=3.43s   p(95)=3.91s
     iterations.....................: 473    15.340525/s
     vus............................: 50     min=50      max=50
     vus_max........................: 50     min=50      max=50
```

After:

```
     http_req_waiting...............: avg=2.76s    min=319.83ms med=2.63s   max=7.05s    p(90)=2.96s   p(95)=4.98s
     http_reqs......................: 525    17.306773/s
     iteration_duration.............: avg=2.8s     min=333.86ms med=2.65s   max=7.08s    p(90)=3.01s   p(95)=5.1s
     iterations.....................: 525    17.306773/s
     vus............................: 50     min=50      max=50
     vus_max........................: 50     min=50      max=50
```

This gives us about `~400ms` of speed up in the p(90) metrics.